backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai-whisper
groq
openai
python-multipart
websockets
pyaudio
torch
numpy
pydub
ffmpeg-python
psutil
slowapi
aiohttp
asyncio
scipy
scikit-learn
pytest
pytest-cov
pytest-asyncio